requirements.txt
Flask==3.0.2
spotipy==2.23.0
python-dotenv==0.21.0
gunicorn==21.2.0
Werkzeug==3.0.0
